Created client side APIs and added authentication to both admin and client and updated Users Test with authentication

+++ src/models/posts.model.js
@@ -1,8 +1,18 @@
+import fs from 'fs';
+import path from 'path';
 import { postDatabase } from './posts.mongo.js';
+import AppError from '../utils/AppError.js';
 
 // Create a Post function
 export async function createPost(data, image) {
-  const post = new postDatabase({ ...data, image });
+  const post = new postDatabase({
+    ...data,
+    image: image
+      ? {
+          url: image,
+        }
+      : undefined,
+  });
   return await post.save();
 }
 
@@ -11,11 +21,17 @@ export async function getAllPosts(skip, limit) {
   return await postDatabase
     .find()
     .sort({
-      createdAt: 1,
+      createdAt: -1, //latest first
     })
     .skip(skip)
     .limit(limit)
     .populate('createdBy')
+    // .populate({
+    //   path: 'category',
+    //   populate: {
+    //     path: 'createdBy',
+    //   },
+    // });
     .populate('category');
 }
 
@@ -23,31 +39,46 @@ export async function getAllPosts(skip, limit) {
 export async function getPostById(id) {
   return await postDatabase.findById(id).populate('createdBy').populate('category');
 }
+
+//Read post by Slug
+export async function getPostBySlug(slug) {
+  const post = await postDatabase.findOne({ slug }).populate('createdBy').populate('category');
+  if (!post) {
+    throw new AppError('Post not found', 404);
+  }
+  return post;
+}
+
 // Update post by ID
 export async function updatePost(id, data, image) {
-  const post = postDatabase.findById(id);
+  const post = await postDatabase.findById(id);
   if (!post) {
     return res.status(404).json({
       message: 'Post not found',
     });
   }
-  if (image) {
-    if (post.image) {
-      const oldImagePath = path.resolve(post.image);
-      fs.unlink(oldImagePath, (err) => {
-        if (err) {
-          console.log('Failed to delete image', err);
-        } else {
-          console.log('Old image deleted:', oldImagePath);
-        }
-      });
-    }
+  if (image && post.image && post.image.url) {
+    const oldImagePath = path.resolve('uploads', 'public', post.image.url);
+    fs.unlink(oldImagePath, (err) => {
+      if (err) {
+        console.log('Failed to delete image', err);
+      } else {
+        console.log('Old image deleted:', oldImagePath);
+      }
+    });
     // post.image = image;
   }
   return await postDatabase
     .findByIdAndUpdate(
       id,
-      { ...data, image },
+      {
+        ...data,
+        image: image
+          ? {
+              url: image,
+            }
+          : undefined,
+      },
       {
         new: true,
       }
@@ -58,11 +89,10 @@ export async function updatePost(id, data, image) {
 
 //Delete post by ID
 export async function deletePost(id) {
-  const fs = require('fs');
-  const post = postDatabase.findById(id);
-  if (post.image) {
-    const imagePath = path.resolve(post.image); // ensure absolute path * check this part in postman
-    fs.unlink(`uploads/${post.image}`, (err) => {
+  const post = await postDatabase.findById(id);
+  if (post && post.image && post.image.url) {
+    const imagePath = path.resolve('uploads', 'public', post.image.url); // ensure absolute path * check this part in postman
+    fs.unlink(imagePath, (err) => {
       if (err) {
         console.error('Error deleting old image:', err);
       }
